move isColourHex to regex-utils
