+++ packages/utils/src/regex-utils/isColourHex.ts
